refactor(desktop): extract user-lock + interrupt prompts
Move per-workspace BrowserUserLockState management, agent-interrupt
confirmation dialog, and the per-WebContents programmatic-input depth
counter into browser-pane/user-lock.ts.

Moved:
- activeUserBrowserLock / releaseUserBrowserLock / ensureUserBrowserLock
- pauseBrowserControlSession
- agentBrowserSessionNeedsInterrupt / confirmBrowserInterrupt
- maybePromptBrowserInterrupt
- isProgrammaticBrowserInput / withProgrammaticBrowserInput

The dedup Set (`userBrowserInterruptPrompts`) and the WeakMap counter
(`programmaticBrowserInputDepth`) are closure-captured by the factory.
Runtime-status awareness is injected as an `isAgentSessionBusy`
callback so the module stays decoupled from the runtime client.

main.ts -207. Typecheck clean.
